# SERVER_STABLE/SERVER_STABLE/openapi.yaml
openapi: 3.0.3
info:
  title: Academy Portal API
  version: 1.0.0
  description: >
    Technical contract for the Academy Portal backend. The API supports public
    course discovery, student enrollments, and admin management of courses,
    teachers, and categories.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: List course categories
      operationId: listCategories
      security: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          example: frontend
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - name
              - -created_at
              - created_at
          example: name
      responses:
        '200':
          description: Paginated categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Categories
      summary: Create a category
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/categories/{categoryId}:
    put:
      tags:
        - Categories
      summary: Update a category
      operationId: updateCategory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Categories
      summary: Delete a category
      operationId: deleteCategory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/admin/audit-logs:
    get:
      tags:
        - Admin
      summary: List audit log entries
      operationId: listAuditLogs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: entity_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - user
              - teacher
              - category
              - course
              - enrollment
              - session
          example: course
        - name: action
          in: query
          required: false
          schema:
            type: string
            enum:
              - create
              - update
              - delete
              - publish
              - archive
              - enroll
              - unenroll
              - login
              - verify_email
              - request_password_reset
              - reset_password
          example: publish
      responses:
        '200':
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/course-daily-stats:
    get:
      tags:
        - Admin
      summary: List course daily enrollment rollups
      operationId: listCourseDailyStats
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: course_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Id'
          example: '9001'
        - name: metric_date
          in: query
          required: false
          schema:
            type: string
            format: date
          example: '2026-05-13'
      responses:
        '200':
          description: Paginated daily rollups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDailyStatListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/jobs/email:
    get:
      tags:
        - Admin
      summary: Inspect email queue counts and recent jobs
      operationId: inspectEmailQueue
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Email queue snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/jobs/maintenance:
    get:
      tags:
        - Admin
      summary: Inspect maintenance queue counts and recent jobs
      operationId: inspectMaintenanceQueue
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Maintenance queue snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/jobs/course-daily-stats:
    post:
      tags:
        - Admin
      summary: Trigger course daily stats refresh
      operationId: triggerCourseDailyStats
      security:
        - BearerAuth: []
      responses:
        '202':
          description: Job queued when Redis workers are enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedJobResponse'
        '200':
          description: Stats synced synchronously when workers are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: >
        Opaque keyset cursor generated from the previous page using
        (created_at, id) or (enrolled_at, id) ordering.
      schema:
        type: string
      example: MjAyNi0wNC0yOFQwOTowMDowMFouMDAwWjo5MDAx
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    CategoryIdParam:
      name: categoryId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
      example: '2001'
...
  responses:
    BadRequest:
      description: Request body, path, or query parameter is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: bad_request
              message: Validation failed for query parameter limit.
              details:
                limit: must be less than or equal to 100
              request_id: req_7f3a0d2e
    Unauthorized:
      description: Missing, expired, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Bearer token is missing or invalid.
              details: {}
              request_id: req_7f3a0d2e
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: You do not have permission to perform this action.
              details:
                required_roles:
                  - admin
              request_id: req_7f3a0d2e
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: Course 9001 was not found.
              details: {}
              request_id: req_7f3a0d2e
    Conflict:
      description: Request conflicts with current system state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: conflict
              message: Student is already enrolled in this course.
              details:
                course_id: '9001'
              request_id: req_7f3a0d2e
    UnprocessableEntity:
      description: Request format is valid JSON but violates semantic validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unprocessable_entity
              message: Course cannot be published while required fields are missing.
              details:
                missing_fields:
                  - description
                  - capacity
              request_id: req_7f3a0d2e
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_server_error
              message: An unexpected error occurred.
              details: {}
              request_id: req_7f3a0d2e
  schemas:
    Id:
      type: string
      pattern: '^[0-9]+$'
      example: '9001'
...
    Role:
      type: string
      enum:
        - student
        - admin
      example: student
...
    CourseStatus:
      type: string
      enum:
        - draft
        - published
        - archived
      example: published
...
    Error:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          example: conflict
        message:
          type: string
          example: Student is already enrolled in this course.
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
          example: req_7f3a0d2e
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    MutationSuccess:
      type: object
      required:
        - ok
        - message
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Resource deleted successfully.
    PaginationMeta:
      type: object
      required:
        - limit
        - has_more
      properties:
        next_cursor:
          type: string
          nullable: true
          example: MjAyNi0wNC0yN1QxMTowMDowMFouMDAwWjo4OTk1
        limit:
          type: integer
          example: 20
        has_more:
          type: boolean
          example: true
    UserSummary:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          format: email
          example: student@academy.dev
        display_name:
          type: string
          nullable: true
          example: Ayan Sarsen
        role:
          $ref: '#/components/schemas/Role'
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-05T10:05:00Z'
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-04-28T09:00:00Z'
...
    Category:
      type: object
      required:
        - id
        - slug
        - name
        - is_active
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Id'
        slug:
          type: string
          example: frontend
        name:
          type: string
          example: Frontend
        description:
          type: string
          nullable: true
          example: Courses focused on client-side web development.
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-04-20T09:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-04-22T09:00:00Z'
    CategoryCreateRequest:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
          example: frontend
        name:
          type: string
          example: Frontend
        description:
          type: string
          nullable: true
          example: Courses focused on client-side web development.
        is_active:
          type: boolean
          default: true
          example: true
    CategoryUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/CategoryCreateRequest'
    CategoryListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AuditLog:
      type: object
      required:
        - id
        - entity_type
        - entity_id
        - action
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Id'
        actor_user_id:
          $ref: '#/components/schemas/Id'
        entity_type:
          type: string
          example: course
        entity_id:
          $ref: '#/components/schemas/Id'
        action:
          type: string
          example: publish
        before_json:
          type: object
          nullable: true
          additionalProperties: true
        after_json:
          type: object
          nullable: true
          additionalProperties: true
        request_id:
          type: string
          nullable: true
          example: req_7f3a0d2e
        ip_address:
          type: string
          nullable: true
          example: 127.0.0.1
        created_at:
          type: string
          format: date-time
          example: '2026-05-13T10:30:00Z'
        actor:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
    AuditLogListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CourseDailyStat:
      type: object
      required:
        - id
        - course_id
        - metric_date
        - active_enrollment_count
        - new_enrollment_count
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Id'
        course_id:
          $ref: '#/components/schemas/Id'
        metric_date:
          type: string
          format: date
          example: '2026-05-13'
        active_enrollment_count:
          type: integer
          example: 42
        new_enrollment_count:
          type: integer
          example: 5
        created_at:
          type: string
          format: date-time
          example: '2026-05-13T00:15:00Z'
        course:
          nullable: true
          type: object
          properties:
            id:
              $ref: '#/components/schemas/Id'
            slug:
              type: string
              example: react-fundamentals
            name:
              type: string
              example: React Fundamentals
            status:
              $ref: '#/components/schemas/CourseStatus'
    CourseDailyStatListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseDailyStat'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    QueueJob:
      type: object
      properties:
        id:
          type: string
          example: '42'
        name:
          type: string
          example: send-email
        attempts_made:
          type: integer
          example: 1
        failed_reason:
          type: string
          nullable: true
          example: null
        timestamp:
          type: string
          format: date-time
          nullable: true
        processed_on:
          type: string
          format: date-time
          nullable: true
        finished_on:
          type: string
          format: date-time
          nullable: true
        data:
          type: object
          additionalProperties: true
    QueueSnapshot:
      type: object
      required:
        - enabled
        - counts
        - jobs
      properties:
        enabled:
          type: boolean
          example: true
        counts:
          type: object
          additionalProperties:
            type: integer
          example:
            waiting: 0
            active: 0
            completed: 12
            failed: 1
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/QueueJob'
    QueuedJobResponse:
      type: object
      required:
        - queued
      properties:
        queued:
          type: boolean
          example: true
        job_id:
          type: string
          nullable: true
          example: 'sync-course-daily-stats-repeat'